requirements-npu.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
datasets
dill
hydra-core
numpy<2.0.0
pandas
peft>=0.15.2
pyarrow>=15.0.0
pybind11
pylatexenc
tensordict>=0.8.0,<=0.10.0,!=0.9.0
transformers==4.52.4
ray==2.46.0
wandb
mathruler
torchdata
einops
qwen_vl_utils
torchvision==0.20.1
